api_drf/viewsets.py
from rest_framework import viewsets

from client.models import Client
from api_drf.serializers import ClientSerializer


class ClientViewSet(viewsets.ModelViewSet):
    queryset = Client.objects.all()
    serializer_class = ClientSerializer
