requirements/local.txt
Django==4.2
dj-database-url==2.0.0
django-cors-headers==4.0.0
django-extensions==3.2.1
django-ninja==0.22.1
djangorestframework==3.14.0
dr-scaffold==2.1.2
psycopg2==2.9.6
python-decouple==3.8
